Rename Reviewable{=> Trait} and introduce new Reviewable interface
