+++ Classes/Mw/Metamorph/Domain/Model/State/ClassMappingContainer.php
@@ -9,9 +9,9 @@ use TYPO3\Flow\Annotations as Flow;
  *
  * @Flow\Scope("prototype")
  */
-class ClassMappingContainer {
+class ClassMappingContainer implements Reviewable {
 
-	use Reviewable;
+	use ReviewableTrait;
 
 	/**
 	 * @var array<\Mw\Metamorph\Domain\Model\State\ClassMapping>
